+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `http` and `grpc` `ServerOptions` to wrap a slice of options into a single element
-
+- Business error middleware
 
 ## 0.1.0 - 2019-09-25